refactor(services): use config flow.protected_branches instead of hardcoded values
Replace hardcoded {"main", "master", "develop"} in task_service.py and
issue_flow_service.py with VibeConfig.get_defaults().flow.protected_branches.

This aligns with the existing pattern used by check_service.py,
flow_service.py, and expired_resource_cleanup_service.py, enabling
custom protected branch configuration through settings.

Changes:
- src/vibe3/services/task_service.py: Add VibeConfig import, use config
- src/vibe3/services/issue_flow_service.py: Add VibeConfig import, use config
- tests: Add tests for custom protected branches from config

Fixes #2040
