+++ src/vibe3/services/task_service.py
@@ -9,6 +9,7 @@ from vibe3.clients.github_client import GitHubClient
 from vibe3.clients.github_labels import GhIssueLabelPort, IssueLabelPort
 from vibe3.clients.protocols import GitHubClientProtocol
 from vibe3.config.orchestra_settings import load_orchestra_config
+from vibe3.config.settings import VibeConfig
 from vibe3.exceptions import InvalidBranchLinkError
 from vibe3.models import IssueState
 from vibe3.models.flow import FlowStatusResponse, IssueLink
@@ -107,11 +108,8 @@ class TaskService:
 
         # Guard: prevent base branches from being linked to issues
         base_branch = self._get_orchestra_config().scene_base_ref.replace("origin/", "")
-        if normalized_branch == base_branch or normalized_branch in (
-            "main",
-            "master",
-            "develop",
-        ):
+        protected_branches = set(VibeConfig.get_defaults().flow.protected_branches)
+        if normalized_branch == base_branch or normalized_branch in protected_branches:
             raise InvalidBranchLinkError(normalized_branch, issue_number)
 
         # Now add the new link
