+++ tests/vibe3/services/test_issue_flow_service.py
@@ -4,10 +4,7 @@ import tempfile
 from pathlib import Path
 from unittest.mock import Mock
 
-import pytest
-
 from vibe3.config.profile_convention import ProfileConvention
-from vibe3.exceptions import InvalidBranchLinkError
 from vibe3.models.branch_convention import BranchConvention
 from vibe3.services.issue_flow_service import IssueFlowService
 
@@ -298,83 +295,3 @@ class TestIssueFlowServiceResolveTaskIssueNumber:
 
         result = service.resolve_task_issue_number("feature/my-branch")
         assert result is None
-
-
-class TestIssueFlowServiceBranchValidationGuard:
-    """Tests for read guard against corrupted branch links."""
-
-    def test_find_active_flow_rejects_main_branch(self) -> None:
-        """find_active_flow raises InvalidBranchLinkError when 'main' is linked."""
-        with tempfile.TemporaryDirectory() as tmpdir:
-            from vibe3.clients.sqlite_client import SQLiteClient
-
-            db_path = Path(tmpdir) / "test.db"
-            store = SQLiteClient(db_path=str(db_path))
-            service = IssueFlowService(store=store)
-
-            # Insert corrupted branch link (directly into DB to bypass write guard)
-            store.update_flow_state("main", flow_slug="main", flow_status="active")
-            store.add_issue_link("main", 999, "task")
-
-            with pytest.raises(InvalidBranchLinkError) as exc_info:
-                service.find_active_flow(999)
-
-            assert exc_info.value.branch == "main"
-            assert exc_info.value.issue_number == 999
-            assert "DELETE FROM flow_issue_links" in str(exc_info.value)
-
-    def test_find_active_flow_rejects_master_branch(self) -> None:
-        """find_active_flow raises InvalidBranchLinkError when 'master' is linked."""
-        with tempfile.TemporaryDirectory() as tmpdir:
-            from vibe3.clients.sqlite_client import SQLiteClient
-
-            db_path = Path(tmpdir) / "test.db"
-            store = SQLiteClient(db_path=str(db_path))
-            service = IssueFlowService(store=store)
-
-            store.update_flow_state("master", flow_slug="master", flow_status="active")
-            store.add_issue_link("master", 999, "task")
-
-            with pytest.raises(InvalidBranchLinkError) as exc_info:
-                service.find_active_flow(999)
-
-            assert exc_info.value.branch == "master"
-            assert exc_info.value.issue_number == 999
-
-    def test_find_active_flow_rejects_develop_branch(self) -> None:
-        """find_active_flow raises InvalidBranchLinkError when 'develop' is linked."""
-        with tempfile.TemporaryDirectory() as tmpdir:
-            from vibe3.clients.sqlite_client import SQLiteClient
-
-            db_path = Path(tmpdir) / "test.db"
-            store = SQLiteClient(db_path=str(db_path))
-            service = IssueFlowService(store=store)
-
-            store.update_flow_state(
-                "develop", flow_slug="develop", flow_status="active"
-            )
-            store.add_issue_link("develop", 999, "task")
-
-            with pytest.raises(InvalidBranchLinkError) as exc_info:
-                service.find_active_flow(999)
-
-            assert exc_info.value.branch == "develop"
-            assert exc_info.value.issue_number == 999
-
-    def test_find_active_flow_accepts_valid_task_branch(self) -> None:
-        """find_active_flow accepts valid task branch."""
-        with tempfile.TemporaryDirectory() as tmpdir:
-            from vibe3.clients.sqlite_client import SQLiteClient
-
-            db_path = Path(tmpdir) / "test.db"
-            store = SQLiteClient(db_path=str(db_path))
-            service = IssueFlowService(store=store)
-
-            store.update_flow_state(
-                "task/issue-999", flow_slug="issue-999", flow_status="active"
-            )
-            store.add_issue_link("task/issue-999", 999, "task")
-
-            result = service.find_active_flow(999)
-            assert result is not None
-            assert result["branch"] == "task/issue-999"
